surfaceflinger: Fix hwcomposer command errors
When using the hwcomposer API, there are two paths that the command may
take. One path results in a direct synchronous IPC message, whereas the
other path will buffer the commands and dispatch later just before
present.

This opens the possibility for out-of order execution of those commands.
and can result in the following errors showing in logcat:

    HwcComposer: command 0x3010000 generated error 3

One scenario where this happens is the destroyLayer() command. This
command is dispatched immediately, but it's possible that the buffered
command stream [that has not yet been sent] contains commands that
reference that layer.

+++ services/surfaceflinger/DisplayHardware/ComposerHal.cpp
@@ -284,6 +284,11 @@ Error Composer::createVirtualDisplay(uint32_t width, uint32_t height,
 
 Error Composer::destroyVirtualDisplay(Display display)
 {
+    Error error = execute();
+    if (error != Error::NONE) {
+        return error;
+    }
+
     auto ret = mClient->destroyVirtualDisplay(display);
     return unwrapRet(ret);
 }
@@ -313,6 +318,11 @@ Error Composer::createLayer(Display display, Layer* outLayer)
 
 Error Composer::destroyLayer(Display display, Layer layer)
 {
+    Error error = execute();
+    if (error != Error::NONE) {
+        return error;
+    }
+
     auto ret = mClient->destroyLayer(display, layer);
     return unwrapRet(ret);
 }
